Add textured mesh rendering and more comments
 + Added assimp library for 3D model loading
 * Changed filename convention from "resource_structures" to simply
"resources"
 * Renamed MessageContents->data to descr
 + Added a void data pointer to MessageContents
 + Added a basic fragment shader in resources/fragment_basic.glsl which
ignores
        lighting, shadows, and texture manipulation
 + Added another overloaded function for BEE::messenger_send()
 + Added initial version of 3D Mesh resource in bee/resources/mesh.cpp
 + Added comments to bee/resources/sprite.cpp
 * Cleaned up Sprite resource handling
 * Made better use of VAOs when drawing Sprites
 * Enabled depth testing when 3D drawing is enabled
 - Removed useless program error messages from BEE::opengl_init()
 * Increased the default camera far plane to 10'000.0 from 1'000.0
 + Added an object data struct to ObjBee to help with storing instanced
data
        ! This implementation is taken from another project of mine so I
will
          polish it up soon
 + Added keys in ObjBee to control the camera in 3D mode
 + Added a new screenshot of the current test room
 - Removed is_hud parameters for the functions in bee/game/draw.cpp in
favor of
	proper utilization of BEE::set_viewport() from
bee/game/room.cpp
 * Fixed indentation formatting in several files
 * Cleaned up add_to_resources() in each primary resource type

+++ bee/game/resources.cpp
@@ -38,7 +38,7 @@ int BEE::free_media() const {
 * @channel: the channel which has finished playback
 */
 void BEE::sound_finished(int channel) {
-	for (int i=0; i<resource_list->sounds.get_amount(); i++) { // Iterate over the sounds in order to remove finished channels from each sound's list
+	for (size_t i=0; i<resource_list->sounds.get_amount(); i++) { // Iterate over the sounds in order to remove finished channels from each sound's list
 		if (get_sound(i) != nullptr) {
 			if (!get_sound(i)->get_is_music()) { // Music cannot be played on multiple channels
 				get_sound(i)->finished(channel); // Remove the finished channel from the list
@@ -60,7 +60,7 @@ double BEE::get_volume() const {
 int BEE::set_volume(double new_volume) {
 	volume = new_volume; // Set the volume
 
-	for (int i=0; i<resource_list->sounds.get_amount(); i++) { // Iterate over the sounds and update them to the new volume
+	for (size_t i=0; i<resource_list->sounds.get_amount(); i++) { // Iterate over the sounds and update them to the new volume
 		if (get_sound(i) != nullptr) {
 			get_sound(i)->update_volume();
 		}
@@ -72,7 +72,7 @@ int BEE::set_volume(double new_volume) {
 * BEE::sound_stop_all() - Immediately stop all sound output
 */
 int BEE::sound_stop_all() const {
-	for (int i=0; i<resource_list->sounds.get_amount(); i++) { // Iterate over the sounds and stop them individually
+	for (size_t i=0; i<resource_list->sounds.get_amount(); i++) { // Iterate over the sounds and stop them individually
 		if (get_sound(i) != nullptr) {
 			get_sound(i)->stop();
 		}
@@ -228,7 +228,7 @@ BEE::Room* BEE::get_room(int id) {
 * @name: the name of the desired object
 */
 BEE::Object* BEE::get_object_by_name(const std::string& name) const {
-	for (int i=0; i<resource_list->objects.get_amount(); i++) { // Iterate over the objects in order to find the first one with the given name
+	for (size_t i=0; i<resource_list->objects.get_amount(); i++) { // Iterate over the objects in order to find the first one with the given name
 		if (get_object(i) != nullptr) {
 			if (get_object(i)->get_name() == name) {
 				return get_object(i); // Return the desired object on success
